ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
networkx==3.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
